bdk: allow Windows builds.
Two repositories required for Windows builds were left out of the
Brillo manifest in order to save space, when we thought that we would
be distributing the entire source tree to developers. Now that this is
no longer the case, we can add them back in to allow adb/fastboot to
build for Windows.

Bug: 28621297
Change-Id: Icd4eb77d5ae2835299c59401899662127456351e
Test: `m -j81 adb`